[MODIFIED prettier]

+++ routes/data/liked.js
@@ -2,20 +2,14 @@ export async function getAll(conn, postId) {
   return conn
     .execute(
       `SELECT li.user, us.nickname, us.profile_image_url FROM LIKED as li join USER as us on li.user = us.id WHERE li.post = ? ORDER BY li.created_at`,
-      [Number(postId)]
+      [Number(postId)],
     )
     .then((result) => result[0]);
 }
 
 export async function create(conn, postId, userId) {
-  return conn.execute('INSERT into LIKED (post, user) values (?, ?)', [
-    Number(postId),
-    userId,
-  ]);
+  return conn.execute('INSERT into LIKED (post, user) values (?, ?)', [Number(postId), userId]);
 }
 export async function remove(conn, postId, userId) {
-  return conn.execute('DELETE FROM LIKED WHERE post = ? and user = ?', [
-    postId,
-    userId,
-  ]);
+  return conn.execute('DELETE FROM LIKED WHERE post = ? and user = ?', [postId, userId]);
 }
